+++ src/__generated__/QueryRenderersBidFlowQuery.graphql.ts
@@ -1,0 +1,116 @@
+/* tslint:disable */
+
+import { ConcreteRequest } from "relay-runtime";
+export type QueryRenderersBidFlowQueryVariables = {
+    readonly saleArtworkID: string;
+};
+export type QueryRenderersBidFlowQueryResponse = {
+    readonly sale_artwork: ({
+    }) | null;
+};
+
+
+
+/*
+query QueryRenderersBidFlowQuery(
+  $saleArtworkID: String!
+) {
+  sale_artwork(id: $saleArtworkID) {
+    ...BidFlow_sale_artwork
+    __id
+  }
+}
+
+fragment BidFlow_sale_artwork on SaleArtwork {
+  bid_increments
+  __id
+}
+*/
+
+const node: ConcreteRequest = (function(){
+var v0 = [
+  {
+    "kind": "LocalArgument",
+    "name": "saleArtworkID",
+    "type": "String!",
+    "defaultValue": null
+  }
+],
+v1 = [
+  {
+    "kind": "Variable",
+    "name": "id",
+    "variableName": "saleArtworkID",
+    "type": "String!"
+  }
+],
+v2 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "__id",
+  "args": null,
+  "storageKey": null
+};
+return {
+  "kind": "Request",
+  "operationKind": "query",
+  "name": "QueryRenderersBidFlowQuery",
+  "id": "0d027936574588c1b968f7d9a87ea74b",
+  "text": null,
+  "metadata": {},
+  "fragment": {
+    "kind": "Fragment",
+    "name": "QueryRenderersBidFlowQuery",
+    "type": "Query",
+    "metadata": null,
+    "argumentDefinitions": v0,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "sale_artwork",
+        "storageKey": null,
+        "args": v1,
+        "concreteType": "SaleArtwork",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "FragmentSpread",
+            "name": "BidFlow_sale_artwork",
+            "args": null
+          },
+          v2
+        ]
+      }
+    ]
+  },
+  "operation": {
+    "kind": "Operation",
+    "name": "QueryRenderersBidFlowQuery",
+    "argumentDefinitions": v0,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "sale_artwork",
+        "storageKey": null,
+        "args": v1,
+        "concreteType": "SaleArtwork",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "bid_increments",
+            "args": null,
+            "storageKey": null
+          },
+          v2
+        ]
+      }
+    ]
+  }
+};
+})();
+(node as any).hash = 'f12416f909511f5de8b83cc919f5d84f';
+export default node;